Add a couple of DNF API methods
/api/v0/dnf/info/<package>,...

Will return a JSON string with the details about the package (or
comma-separated list of packages).

/api/v0/dnf/transaction/<package>,...

Will return a JSON string with the list of packages that would be
installed to satisfy the package or comma-separated list of packages.

This code isolates the DNF Base object in its own thread. Communication
is via the dnf_q QueueFactory Queue, and results are returned via a
different queue passed as part of the query.