Add API methods for status and reading logs
This adds /api/v0/compose/status and /api/v0/compose/log/<kbytes>, see
README.demo for details.

It also adds a stub for /api/v0/compose/cancel, but that is not
implemented yet.

It also changes the default URLs to http instead of https (proxies don't
work with https), and it explicitly sets the --proxy on the kickstart
url and repo lines because anaconda-24.13.7 isn't using the global
--proxy for repos, only url the url URL.