Add /api/v0/compose/types to return supported compose types
Returns a JSON object with the supported types and a bool flag
indicating where or not that type can actually be used. For the demo
code these are all False, except for creating an iso.

Also adds the ability to pass a compose 'type' to the /api/v0/compose
call to demonstrate failure of the call if it isn't set to 'iso'.