Add a bill of materials file to the built image.
This goes in /root, just like anaconda-ks.cfg would.  For now it's just
the list of all packages that went into the image, in json format.
Other things could go into it pretty easily.  Note that I am running
this as a non-chrooted %post script because we cannot assume python will
be in the image.