Convert the recipe to JSON before sending it via the API
Seriously, leaving the recipe as TOML embedded in JSON to be
double-processed on the javascript side is just making things harder for
ourselves. The backend has access to better TOML parsers, so just go
ahead and convert the web data to the web language.

While I was at it, rearrange some of the data slightly: change the
"name" property in the TOML data to "description" (the header is already
the recipe name), and when converting to JSON convert the header name to
a "name" property.